Gym-Management-System/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
email-validator==2.1.0
python-dateutil==2.8.2
python-jose[cryptography]==3.3.0
argon2-cffi==23.1.0